Sentinel_2/requirements.txt
affine==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
eval_type_backport==0.3.0
filelock==3.19.1
folium==0.20.0
fsspec==2025.9.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.45
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
phidown==0.1.23
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
s5cmd==0.3.3
sentry-sdk==2.46.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
xyzservices==2025.11.0
zipp==3.23.0
